refactor(server): move client perf rows out of db.ts and instrument the shed ladder
The desktop_shell column's four lines pushed server/db.ts past its zero-slack
monolith ceiling, and the ratchet's answer is extraction: the client perf
report insert and retention prune move whole to client_perf_reports_db.ts
beside the schema module that already owns the DDL (pure re-export kept in
db.ts, so no caller or test double re-points), and the ceiling is lowered to
the new size.

Riding with it, from review: a woc_client_raw_summary_shed_total counter by
deepest rung reached, the only fleet-wide readout of how far the ladder goes;
ingest, reporter and metrics pins for the contracts the review found
untested (posted markers stripped, the client's rawSummary keys locked to the
server's known set, the runtime label bounded to exactly two families); and
the DEPLOY notes on the series identity change and the fps_avg step.
